backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
alembic==1.10.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-docx==0.8.11
python-pptx==0.6.21
PyJWT==2.8.0
requests==2.31.0
openai==0.27.8
email-validator==1.3.1